Fixed non-showing loading indicator in processForm
"hidePopup" was called right after creating the AJAX request, but it should be called after the request finished e.g. in the method that handles the readyState 4. Otherwise the indicator would show, then the request created and the indicator hidden which leads to not showing the indicator at all in e.g. Chromium 48.
I noticed this bug when deleting huge directories which showed me no loading indicator.